Moved some parameters from size_t to unsigned int.
git-svn-id: svn+ssh://svn.tuxfamily.org/svnroot/scengine/scengine/trunk@234 d5589c5c-db5c-46e9-ada4-dbbd3b4b0917

+++ include/SCE/utils/SCEList.h
@@ -132,7 +132,8 @@ SCE_SListIterator* SCE_List_LocateIterator (const SCE_SList*, void*,
                                             SCE_FListCompareData);
 unsigned int SCE_List_LocateIndex (const SCE_SList*, void*,
                                    SCE_FListCompareData);
-void SCE_List_QuickSortRange (SCE_SList*, size_t, size_t, SCE_FListCompareData);
+void SCE_List_QuickSortRange (SCE_SList*, unsigned int, unsigned int,
+                              SCE_FListCompareData);
 void SCE_List_QuickSort (SCE_SList*, SCE_FListCompareData);
 
 /**
